Auth

+++ tema07/proyectos/03-mvc-geslibros/controllers/libro.php
@@ -3,7 +3,7 @@
 class Libro extends Controller
 {
 
-    function __construct()
+    public function __construct()
     {
 
         parent::__construct();
@@ -20,6 +20,9 @@ class Libro extends Controller
     {
         // inicio o continuo la sesión
         session_start();
+
+        // Creo un token CSRF
+        $_SESSION['csrf_token'] = bin2hex(random_bytes(32));
 
         // Compruebo si hay mensajes de éxito
         if (isset($_SESSION['success'])) {
@@ -55,7 +58,7 @@ class Libro extends Controller
         session_start();
 
         // Creo un token para evitar ataques CSRF
-        $_SESSION['csrf_token'] = bin2hex(random_bytes(32));    
+        $_SESSION['csrf_token'] = bin2hex(random_bytes(32));
 
         // Creo un objeto vacío de la clase libro
         $this->view->libro = new classLibro();
@@ -253,7 +256,7 @@ class Libro extends Controller
         @param int $id: id del alumno a editar
 
     */
-    function editar($param = [])
+    public function editar($param = [])
     {
 
         # obtengo el id del alumno que voy a editar
@@ -262,33 +265,31 @@ class Libro extends Controller
         // -- edit es el nombre del método
         // -- $param es un array porque puedo pasar varios parámetros a un método
 
-        // inicio o continuo la sesion
         session_start();
 
-        # asigno id a una propiedad de la vista
         $this->view->id = htmlspecialchars($param[0]);
 
         # obtengo el token CSRF
         $this->view->csrf_token = $param[1];
 
-        # Validacion CSRF
+        // Validacion CSRF
         if (!hash_equals($_SESSION['csrf_token'], $this->view->csrf_token)) {
             require_once 'controllers/error.php';
-            $controller = new Errores('Peticion no válida');
+            $controller = new Errores();
             exit();
         }
 
-        # obtengo el objeto de la clase libro con el id asociado
+        # obtener objeto de la clase alumno con el id asociado
         $this->view->libro = $this->model->read($this->view->id);
 
-        # Compruebo si hay errores de validación
-        if(isset($_SESSION['error'])) {
+        // Compruebo si hay errores de validación
+        if (isset($_SESSION['error'])) {
 
             // Creo la propiedad error de la vista
             $this->view->error = $_SESSION['error'];
 
-            // Creo la propiedad libro de la vista
-            $this->view->libro = $_SESSION['libro'];
+            // Creo la propiedad alumno de la vista
+            $this->view->alumno = $_SESSION['libro'];
 
             // Creo la propiedad mensaje de error
             $this->view->mensaje = 'Error en el formulario';
@@ -297,8 +298,7 @@ class Libro extends Controller
             unset($_SESSION['libro']);
 
             // Elimino la variable de sesión error
-            unset($_SESSION['errores']);
-
+            unset($_SESSION['error']);
         }
 
         # title
@@ -306,7 +306,7 @@ class Libro extends Controller
 
         # obtener objeto de la clase alumno con el id pasado
         // Necesito crear el método read en el modelo
-        $this->view->libro = $this->model->read($id);
+        $this->view->libro = $this->model->read($this->view->id);
 
         // Creo la propiedad autores en la vista
         $this->view->autores = $this->model->get_autores();
@@ -335,19 +335,18 @@ class Libro extends Controller
     public function update($param = [])
     {
 
-        # inicio o continuo la sesion
         session_start();
 
-        # obtengo el id del alumno que voy a editar
+        # Cargo id del alumno
         $id = htmlspecialchars($param[0]);
 
-        # obtengo el token CSRF
+        # Obtetngo el token CSRF
         $csrf_token = $param[1];
 
-        # Validacion CSRF
+        // Validación CSRF
         if (!hash_equals($_SESSION['csrf_token'], $csrf_token)) {
             require_once 'controllers/error.php';
-            $controller = new Errores('Peticion no válida');
+            $controller = new Errores();
             exit();
         }
 
@@ -364,7 +363,7 @@ class Libro extends Controller
         # Con los detalles formulario creo objeto alumno
         $libro_form = new classLibro(
 
-            $id,
+            null,
             $titulo,
             $precio,
             $stock,
@@ -376,155 +375,135 @@ class Libro extends Controller
 
         );
 
-        # Actualizo base  de datos
-        # Obtengo los detalles del libro de la base de datos
-        $libro = $this->model->read($libro_form);
-
-        # Validacion de los datos
-        # Validacion en caso de que haya sufrido modificaciones en el campo correspondiente
+        // Validacion de datos
+        // Valido en caso de que haya sufrido modificaciones en el campo correspondiente
         $error = [];
 
-        # Control de cambios en los campos
+        // Control de cambiosen los campos
         $cambios = false;
 
-        # Validacion del título
-        # Reglas: obligatorio
-        if(strcmp($titulo, $libro->titulo != 0)) {
+        // Validación del título
+        // Reglas: obligatorio
+        if (empty($titulo)) {
+            $error['titulo'] = 'El título es obligatorio';
+        } elseif ($titulo != $this->model->read($id)->titulo) {
             $cambios = true;
-            if (empty($titulo)) {
-                $error['titulo'] = 'El título es obligatorio';
-            }
         }
 
-        # Validacion del precio
-        # Reglas: obligatorio, numérico
-        if(strcmp($precio, $libro->precio != 0)) {
+        // Validación del precio
+        // Reglas: obligatorio, numérico
+        if (empty($precio)) {
+            $error['precio'] = 'El precio es obligatorio';
+        } elseif (!is_numeric($precio)) {
+            $error['precio'] = 'El precio debe ser numérico';
+        } elseif ($precio != $this->model->read($id)->precio) {
             $cambios = true;
-            if (empty($precio)) {
-                $error['precio'] = 'El precio es obligatorio';
-            } elseif (!is_numeric($precio)) {
-                $error['precio'] = 'El precio debe ser numérico';
-            }
         }
 
-        # Validacion del stock
-        # Reglas: obligatorio, numérico
-        if(strcmp($stock, $libro->stock != 0)) {
+        // Validación del stock
+        // Reglas: obligatorio, numérico
+        if (empty($stock)) {
+            $error['stock'] = 'El stock es obligatorio';
+        } elseif (!is_numeric($stock)) {
+            $error['stock'] = 'El stock debe ser numérico';
+        } elseif ($stock != $this->model->read($id)->stock) {
             $cambios = true;
-            if (empty($stock)) {
-                $error['stock'] = 'El stock es obligatorio';
-            } elseif (!is_numeric($stock)) {
-                $error['stock'] = 'El stock debe ser numérico';
-            }
         }
 
-        # Validacion de la fecha de edición
-        # Reglas: obligatorio
-        if(strcmp($fecha_edicion, $libro->fecha_edicion != 0)) {
+        // Validación de la fecha de edición
+        // Reglas: obligatorio
+        if (empty($fecha_edicion)) {
+            $error['fecha_edicion'] = 'La fecha de edición es obligatoria';
+        } elseif ($fecha_edicion != $this->model->read($id)->fecha_edicion) {
             $cambios = true;
-            if (empty($fecha_edicion)) {
-                $error['fecha_edicion'] = 'La fecha de edición es obligatoria';
-            }
         }
 
-        # Validacion del ISBN
-        # Reglas: obligatorio, formato ISBN y clave secundaria
-        if(strcmp($isbn, $libro->isbn != 0)) {
+        // Validación del ISBN
+        // Reglas: obligatorio, formato ISBN y clave secundaria
+        if (strcmp($isbn, $libro_form->isbn) != 0) {
             $cambios = true;
-            // Expresión regular para ISBN
-            // Formato: 3-598-21508-8
+            // Expresión regular para validar el DNI
+            // 8 números seguidos de una letra
             $options = [
                 'options' => [
-                    'regexp' => '/^\d{1,5}-\d{1,7}-\d{1,6}-\d{1,1}$/'
+                    'rexexp' => '/^\d{1,5}-\d{1,7}-\d{1,6}-\d{1,1}$/'
                 ]
             ];
 
             if (empty($isbn)) {
                 $error['isbn'] = 'El ISBN es obligatorio';
-            } elseif (!filter_var($isbn, FILTER_VALIDATE_REGEXP, $options)) {
-                $error['isbn'] = 'El ISBN no es válido';
+            } else if (!filter_var($isbn, FILTER_VALIDATE_REGEXP, $options)) {
+                $error['isbn'] = 'Formato ISBN no es correcto';
             }
         }
 
-        # Validacion del id_autor
-        # Reglas: obligatorio, entero, clave ajena
-        if ($id_autor =! $libro->autor_id) {
+        // Validación del id_autor
+        // Reglas: obligatorio, numérico, clave ajena
+        if (empty($id_autor)) {
+            $error['id_autor'] = 'El autor es obligatorio';
+        } elseif ($id_autor != $this->model->read($id)->id_autor) {
             $cambios = true;
-            if (empty($id_autor)) {
-                $error['id_curso'] = 'El curso es obligatorio';
-            } else if (!filter_var($id_autor, FILTER_VALIDATE_INT)) {
-                $error['id_curso'] = 'El formato del curso no es correcto';
-            } else if (!$this->model->validateForeignKeyCurso($id_autor)) {
-                $error['id_curso'] = 'El curso no existe';
-            }
         }
 
-        # Validacion del id_editorial
-        # Reglas: obligatorio, entero, clave ajena
-        if ($id_editorial =! $libro->editorial_id) {
+        // Validación del id_editorial
+        // Reglas: obligatorio, numérico, clave ajena
+        if (empty($id_editorial)) {
+            $error['id_editorial'] = 'La editorial es obligatoria';
+        } elseif ($id_editorial != $this->model->read($id)->id_editorial) {
             $cambios = true;
-            if (empty($id_editorial)) {
-                $error['id_editorial'] = 'La editorial es obligatoria';
-            } else if (!filter_var($id_editorial, FILTER_VALIDATE_INT)) {
-                $error['id_editorial'] = 'El formato de la editorial no es correcto';
-            } else if (!$this->model->validateForeignKeyEditorial($id_editorial)) {
-                $error['id_editorial'] = 'La editorial no existe';
-            }
         }
 
-        # Validacion del id_generos
-        # Reglas: obligatorio, entero, clave ajena, array
-        if ($id_generos =! $libro->generos) {
-            $cambios = true;
-            if (empty($id_generos)) {
-                $error['id_generos'] = 'El género es obligatorio';
-            } else if (!is_array($id_generos)) {
-                $error['id_generos'] = 'El género debe ser un número entero';
-            } else {
-                foreach ($id_generos as $genero) {
-                    if (!is_numeric($genero)) {
-                        $error['id_generos'] = 'El género debe ser un número entero';
-                    }
+        // Validación del id_generos
+        // Reglas: obligatorio, numérico, clave ajena
+        if (empty($id_generos)) {
+            $error['id_generos'] = 'El género es obligatorio';
+        } elseif (!is_array($id_generos)) {
+            $error['id_generos'] = 'El género debe ser un número entero';
+        } else {
+            foreach ($id_generos as $genero) {
+                if (!is_numeric($genero)) {
+                    $error['id_generos'] = 'El género debe ser un número entero';
                 }
             }
         }
 
-        # Si hay errores 
+        // Si hay errores
         if (!empty($error)) {
 
-            # Creo la variable de sesión error
-            $_SESSION['error'] = $error;
+            // Formulario no ha sido validado
+            // Tengo que redireccionar al formulario de nuevo
 
-            # Creo la variable de sesión libro
+            // Creo la variable de sessión alumno con los datos del formulario
             $_SESSION['libro'] = $libro_form;
 
-            # Redirecciono al formulario de editar
-            header('location:' . URL . 'libro/editar/' . $id . '/' . $csrf_token);
+            // Creo la variable de sessión error con los errores
+            $_SESSION['error'] = $error;
 
-            # Salgo del script
-            exit;
+            // redireciona al formulario de nuevo
+            header('location:' . URL . 'libro/editar/' . $id);
+            exit();
         }
 
-        # Compruebo si hay cambios
-        if ($cambios) {
-            # Genero un mensaje de éxito
+        // Compruebo si ha habido cambios
+        if (!$cambios) {
+            // Genero mensaje de éxito
             $_SESSION['mensaje'] = 'No se han realizado cambios';
 
-            # Redirecciono al main de libro
+            // redireciona al main de alumno
             header('location:' . URL . 'libro');
             exit();
         }
 
-        # Necesito crear el método update en el modelo
+        # Actualizo base  de datos
+        // Necesito crear el método update en el modelo
         $this->model->update($libro_form, $id);
 
-        # Genero mensaje de exito
-        $_SESSION['mensaje'] = 'Libro actualizado con éxito';
+        // Genero mensaje de éxito
+        $_SESSION['mensaje'] = 'Alumno actualizado con éxito';
 
-        # Cargo el controlador principal de libro
-        header('location' . URL . 'libro');
+        # Cargo el controlador principal de alumno
+        header('location:' . URL . 'libro');
         exit();
-
     }
 
     /*
@@ -534,21 +513,41 @@ class Libro extends Controller
 
         url asociada: /alumno/eliminar/id
 
-        @param
+        @param1
             :int $id: id del alumno a eliminar
     */
     public function eliminar($param = [])
     {
+        session_start();
 
-        # Cargo id del alumno
-        $id = $param[0];
+        // Obtengo el ID y el token CSRF
+        $id = htmlspecialchars($param[0]);
 
-        # Elimino alumno de la base de datos
-        // Necesito crear el método delete en el modelo
+        // obtengo el tokenn CSRF
+        $csrf_token = $param[1];
+
+        // Validar CSRF
+        if (!hash_equals($_SESSION['csrf_token'], $csrf_token)) {
+            require_once 'controllers/error.php';
+            $controller = new Errores();
+            exit();
+        }
+
+        // Validar ID del libro
+        if (!$this->model->validateIdLibro($id)) {
+            $_SESSION['error'] = 'ID no válido';
+            header('location:' . URL . 'libro');
+            exit();
+        }
+
+        // Eliminar el libro
         $this->model->delete($id);
 
-        # Cargo el controlador principal de alumno
+        // Generar mensaje de éxito
+        $_SESSION['mensaje'] = 'Libro eliminado con éxito';
+
         header('location:' . URL . 'libro');
+        
     }
 
     /*
@@ -563,9 +562,28 @@ class Libro extends Controller
     */
     public function mostrar($param = [])
     {
+        # inicio o continuo la sesión
+        session_start();
 
         # Cargo id del alumno
-        $id = $param[0];
+        $id = htmlspecialchars($param[0]);
+
+        # obtengo el token CSRF
+        $csrf_token = $param[1];
+
+        // Validación CSRF
+        if (!hash_equals($_SESSION['csrf_token'], $csrf_token)) {
+            require_once 'controllers/error.php';
+            $controller = new Errores();
+            exit();
+        }
+
+        // Validar ID del libro
+        if (!$this->model->validateIdLibro($id)) {
+            $_SESSION['error'] = 'ID no válido';
+            header('location:' . URL . 'libro');
+            exit();
+        }
 
         # Cargo el título
         $this->view->title = "Mostrar - Gestión de Alumnos";
@@ -573,8 +591,9 @@ class Libro extends Controller
         # Obtengo los detalles del alumno mediante el método read del modelo
         $this->view->libro = $this->model->read($id);
 
-        # obtener los cursos
-        $this->view->cursos = $this->model->get_cursos();
+        $this->view->autores = $this->model->get_autores();
+
+        $this->view->editoriales = $this->model->get_editoriales();
 
         # Cargo la vista
         $this->view->render('libro/mostrar/index');
@@ -596,8 +615,22 @@ class Libro extends Controller
     public function filtrar()
     {
 
+        // inicio o continuo la sesión
+        session_start();
+
         # Obtengo la expresión de búsqueda
-        $expresion = $_GET['expresion'];
+        $expresion = htmlspecialchars($_GET['expresion']);
+
+        // obtengo el token CSRF
+        $csrf_token = htmlspecialchars($_GET['csrf_token']);
+
+        // Validación CSRF
+        if (!hash_equals($_SESSION['csrf_token'], $csrf_token)) {
+            require_once 'controllers/error.php';
+            $controller = new Errores();
+            exit();
+        }
+
 
         # Cargo el título
         $this->view->title = "Filtrar por: {$expresion} - Gestión de Libros";
@@ -624,6 +657,22 @@ class Libro extends Controller
     public function ordenar($param = [])
     {
 
+        // inicio o continuo la sesión
+        session_start();
+
+        // Obtener criterio
+        $id = (int) htmlspecialchars($param[0]);
+
+        // Obtenerr csrf_token
+        $csrf_token = $param[1];
+
+        // Validación CSRF
+        if (!hash_equals($_SESSION['csrf_token'], $csrf_token)) {
+            require_once 'controllers/error.php';
+            $controller = new Errores();
+            exit();
+        }
+
         # Criterios de ordenación
         $criterios = [
             1 => 'Id',
@@ -636,9 +685,6 @@ class Libro extends Controller
             8 => 'Editorial',
             9 => 'Generos'
         ];
-
-        # Obtengo el id del campo por el que se ordenarán los alumnos
-        $id = $param[0];
 
 
         # Cargo el título
